Fix: mat1/mat2 dtype mismatch in _safe_decoder_forward
Root cause: ref_for_layer.float() and query_pos_head FP16 weights mismatch

Fixes:
1. query_pos_head: Cast input to weight dtype before forward
2. layer_norm: Remove .float() from output

Fixes: mat1 and mat2 must have the same dtype (Float vs Half)

+++ backend/hsg_detr/nn/sparse_global_token.py
@@ -880,10 +880,10 @@ class RTDETRDecoderSGB(RTDETRDecoder):
                 shapes,
                 padding_mask,
                 attn_mask,
-                self.query_pos_head(ref_for_layer).to(dtype=output.dtype),
+                self.query_pos_head(ref_for_layer.to(dtype=self.query_pos_head.weight.dtype)).to(dtype=output.dtype),
             )
 
-            head_input = F.layer_norm(output.float(), (output.shape[-1],))
+            head_input = F.layer_norm(output, (output.shape[-1],))
             bbox = self.dec_bbox_head[i](head_input)
             bbox = self.BBOX_DELTA_LIMIT * torch.tanh(bbox / self.BBOX_DELTA_LIMIT)
             refined_bbox = torch.sigmoid(bbox + _safe_inverse_sigmoid(ref_for_layer, eps=self.REFINE_EPS))
